requirements.txt

requests==2.31.0
onetimepass==1.0.1
python-multipart
pyotp
fastapi
uvicorn
python-dotenv
kiteconnect
jinja2
